requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.46
pydantic==1.10.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
email-validator==1.3.1
psycopg2-binary==2.9.8